+++ src/core/iPath.Application/Localization/LocalizationSettings.cs
@@ -4,8 +4,6 @@ public class LocalizationSettings
 {
     public const string ConfigName = "LocalizationSettings";
 
-    public bool Active { get; set; }
-    public bool AddMissingStrings { get; set; }
     public bool AutoSave { get; set; }
 
     /// <summary>